% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractmeteobr-package.R
\docType{package}
\name{extractmeteobr-package}
\alias{extractmeteobr}
\alias{extractmeteobr-package}
\title{extractmeteobr: area-weighted municipal meteorological time series}
\description{
Converts gridded BR-DWGD meteorological NetCDF rasters into area-weighted
time series for Brazilian municipalities and states.
}
\author{
\strong{Maintainer}: LHMET \email{lhmet@ufsm.br}

Authors:
\itemize{
  \item Jônatan Dupont Tatsch (\href{https://orcid.org/0000-0003-0930-9461}{ORCID}) (Contact: jdtatsch [at] gmail.com)
}

}
\keyword{internal}
